don't create symlink when existing

+++ install.sh
@@ -15,7 +15,9 @@ nix shell 'nixpkgs#acl' --command bash -c 'sudo env PATH=$PATH setfacl -k /tmp'
 # POSIX way to get script's dir: https://stackoverflow.com/a/29834779/12156188
 script_dir="$(cd -P -- "$(dirname -- "$(command -v -- "$0")")" && pwd -P)"
 
-ln -s "${script_dir}/nix" /home/vscode/.config/home-manager
+if [ ! -e /home/vscode/.config/home-manager ]; then
+  ln -s "${script_dir}/nix" /home/vscode/.config/home-manager
+fi
 
 nix run '/home/vscode/.config/home-manager/#homeConfigurations."vscode".activationPackage'
